map + modif matrix,point,charmachine
Adt matrix aku ubah elemennya dari integer ke char karena isinya matriks emg char kan nama

adt point aku ubah elemennya dari float ke int, karena lokasinya pasti int.

charmachine aku modif biar bisa membuka txtfile dengan namafile yang kita mau

+++ DaftarADT/point.c
@@ -7,7 +7,7 @@
 #include "point.h"
 #include <math.h>
 
-POINT MakePOINT (float X, float Y){
+POINT MakePOINT (int X, int Y){
     POINT p ;
     Absis(p) = X ;
     Ordinat(p) = Y;
@@ -16,7 +16,7 @@ POINT MakePOINT (float X, float Y){
 
 void BacaPOINT (POINT * P){
     float x, y;
-    scanf("%f %f", &x, &y);
+    scanf("%d %d", &x, &y);
     Absis(*P) = x ;
     Ordinat(*P) = y;
 }
